new searches activity

+++ app/src/main/java/wichura/de/camperapp/models/SearchItem.java
@@ -2,12 +2,13 @@ package wichura.de.camperapp.models;
 
 /**
  * Created by bwichura on 07.02.2017.
+ *
  */
 
 public class SearchItem {
 
 
-    private Integer id;
+    private Long id;
 
     private String description;
 
@@ -23,11 +24,11 @@ public class SearchItem {
 
     private Integer resultCount;
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
